qcacld-3.0: Add support for y2038 Compliance
As a part of y2038 fixes, all the time related data types
should be updated to 64-bit.

Fix Y2038 issue in qdf_mc_timer_get_system_time() by
avoiding the use of 32-bit time representations that
overflow.

+++ core/sme/src/csr/csr_api_roam.c
@@ -1629,7 +1629,7 @@ static void csr_tsm_stats_rsp_processor(struct mac_context *mac, void *pMsg)
 static void csr_send_ese_adjacent_ap_rep_ind(struct mac_context *mac,
 					struct csr_roam_session *pSession)
 {
-	uint32_t roamTS2 = 0;
+	uint64_t roamTS2 = 0;
 	struct csr_roam_info *roam_info;
 	struct pe_session *pe_session = NULL;
 	uint8_t sessionId = WLAN_UMAC_VDEV_ID_MAX;
